requirements.txt
faiss-cpu==1.7.4
fastapi==0.108.0
langchain
langchain-community
langchain-core
langchain-openai
langchainhub
langsmith
numpy==1.26.3
openai
pandas==2.1.4
posthog==3.3.0
pydantic==2.5.3
pydantic_core==2.14.6
pypdf==4.0.0
python-dotenv==1.0.0
requests==2.31.0
uvicorn==0.25.0
mysql-connector-python
langdetect
beautifulsoup4==4.12.2
mysqlclient
python-multipart
deep-translator
reportlab
semantic-router
langdetect